Renamed files for `JoltBodyImpl3D`
