+++ src/joints/jolt_joint_impl_3d.cpp
@@ -1,6 +1,6 @@
 #include "jolt_joint_impl_3d.hpp"
 
-#include "objects/jolt_body_3d.hpp"
+#include "objects/jolt_body_impl_3d.hpp"
 #include "spaces/jolt_space_3d.hpp"
 
 namespace {
