+++ src/objects/jolt_physics_direct_body_state_3d.cpp
@@ -1,6 +1,6 @@
 #include "jolt_physics_direct_body_state_3d.hpp"
 
-#include "objects/jolt_body_3d.hpp"
+#include "objects/jolt_body_impl_3d.hpp"
 #include "spaces/jolt_physics_direct_space_state_3d.hpp"
 #include "spaces/jolt_space_3d.hpp"
