+++ src/servers/jolt_physics_server_3d.cpp
@@ -7,7 +7,7 @@
 #include "joints/jolt_pin_joint_impl_3d.hpp"
 #include "joints/jolt_slider_joint_impl_3d.hpp"
 #include "objects/jolt_area_impl_3d.hpp"
-#include "objects/jolt_body_3d.hpp"
+#include "objects/jolt_body_impl_3d.hpp"
 #include "servers/jolt_project_settings.hpp"
 #include "shapes/jolt_shape_3d.hpp"
 #include "spaces/jolt_job_system.hpp"
