+++ src/spaces/jolt_contact_listener_3d.cpp
@@ -1,7 +1,7 @@
 #include "jolt_contact_listener_3d.hpp"
 
 #include "objects/jolt_area_impl_3d.hpp"
-#include "objects/jolt_body_3d.hpp"
+#include "objects/jolt_body_impl_3d.hpp"
 #include "spaces/jolt_space_3d.hpp"
 
 namespace {
