+++ src/spaces/jolt_motion_filter_3d.cpp
@@ -1,6 +1,6 @@
 #include "jolt_motion_filter_3d.hpp"
 
-#include "objects/jolt_body_3d.hpp"
+#include "objects/jolt_body_impl_3d.hpp"
 #include "objects/jolt_collision_object_3d.hpp"
 #include "servers/jolt_physics_server_3d.hpp"
 #include "shapes/jolt_motion_shape.hpp"
