+++ src/spaces/jolt_physics_direct_space_state_3d.cpp
@@ -1,7 +1,7 @@
 #include "jolt_physics_direct_space_state_3d.hpp"
 
 #include "objects/jolt_area_impl_3d.hpp"
-#include "objects/jolt_body_3d.hpp"
+#include "objects/jolt_body_impl_3d.hpp"
 #include "objects/jolt_collision_object_3d.hpp"
 #include "servers/jolt_physics_server_3d.hpp"
 #include "servers/jolt_project_settings.hpp"
