+++ src/spaces/jolt_space_3d.cpp
@@ -2,7 +2,7 @@
 
 #include "joints/jolt_joint_impl_3d.hpp"
 #include "objects/jolt_area_impl_3d.hpp"
-#include "objects/jolt_body_3d.hpp"
+#include "objects/jolt_body_impl_3d.hpp"
 #include "servers/jolt_project_settings.hpp"
 #include "spaces/jolt_contact_listener_3d.hpp"
 #include "spaces/jolt_layer_mapper.hpp"
